fixed bugs

+++ app/Http/Controllers/AdminPostsController.php
@@ -20,16 +20,9 @@ class AdminPostsController extends Controller
      */
     public function index()
     {
-        //
-
-
-        $posts = Post::all();
-
-
-
+        $posts = Post::paginate(6);
+        $categories = Category::all();
         return view('admin.posts.index', compact('posts','categories'));
-
-
     }
 
     /**
@@ -39,12 +32,7 @@ class AdminPostsController extends Controller
      */
     public function create()
     {
-        //
-
-
         $categories = Category::lists('name','id')->all();
-
-
         return view('admin.posts.create', compact('categories'));
     }
 
@@ -56,41 +44,18 @@ class AdminPostsController extends Controller
      */
     public function store(PostsCreateRequest $request)
     {
-        //
-
         $input = $request->all();
-
-
         $user = Auth::user();
 
-
         if($file = $request->file('photo_id')){
-
-
             $name = time() . $file->getClientOriginalName();
-
-
             $file->move('images', $name);
-
             $photo = Photo::create(['file'=>$name]);
-
-
             $input['photo_id'] = $photo->id;
-
-
         }
 
-
         $user->posts()->create($input);
-
-
         return redirect('/admin/posts');
-
-
-
-
-
-
     }
 
     /**
@@ -101,7 +66,9 @@ class AdminPostsController extends Controller
      */
     public function show($id)
     {
-        //
+        $post = Post::findOrFail($id);
+        $comments = $post->comments()->whereIsActive(1)->get();
+        return view('post', compact('post','comments'));
     }
 
     /**
@@ -112,15 +79,9 @@ class AdminPostsController extends Controller
      */
     public function edit($id)
     {
-        //
-
         $post = Post::findOrFail($id);
-
         $categories = Category::lists('name','id')->all();
-
         return view('admin.posts.edit', compact('post','categories'));
-
-
     }
 
     /**
@@ -132,37 +93,17 @@ class AdminPostsController extends Controller
      */
     public function update(Request $request, $id)
     {
-        //
-
         $input = $request->all();
 
-
-
         if($file = $request->file('photo_id')){
-
-
             $name = time() . $file->getClientOriginalName();
-
-
             $file->move('images', $name);
-
             $photo = Photo::create(['file'=>$name]);
-
-
             $input['photo_id'] = $photo->id;
-
-
         }
 
-
-      Auth::user()->posts()->whereId($id)->first()->update($input);
-
-
+        Auth::user()->posts()->whereId($id)->first()->update($input);
         return redirect('/admin/posts');
-
-
-
-
     }
 
     /**
@@ -173,16 +114,19 @@ class AdminPostsController extends Controller
      */
     public function destroy($id)
     {
-        //
-
         $post = Post::findOrFail($id);
-
-        unlink(public_path() . $post->photo->file);
-
+        if($post->photo)
+            unlink(public_path() . $post->photo->file);
         $post->delete();
-
         return redirect('/admin/posts');
-
-
     }
+
+
+    public function post($id)
+    {
+        $post = Post::findOrFail($id);
+        $comments = $post->comments()->whereIsActive(1)->get();
+        return view('post', compact('post','comments'));
+    }
+
 }
